Refactoring package
