+++ modules/gitbucket-scala-client/src/main/scala/com/github/virtualstack/gitbucket/client/Repository.scala
