+++ modules/gitbucket-scala-client/src/main/scala/com/github/virtualstack/gitbucket/client/services/HttpService.scala
@@ -1,4 +1,4 @@
-package com.github.virtualstack.gitbucket.client
+package com.github.virtualstack.gitbucket.client.services
 
 trait HttpService {
   val httpClient: HttpClient = HttpClient
